Add an "Open idea" (#256)
I realised that we never noted this explicitly in newer editions as the
programme has went on, so this PR adds some placeholder text with the
necessary format. Of course, such a project if proposed would be
selected based on a mentor's availability and discretion. This makes
external contributors aware that there is such a possibility.